requirements.txt
# Core Deep Learning
torch>=2.5.0
torchvision
timm
diffusers
transformers
accelerate
huggingface-hub>=0.25
safetensors

# Image Processing
opencv-python-headless  # use headless for server environments
scikit-image
kornia
einops
pillow
image_proc  # assuming this is your own module or local package

# Backend / API
fastapi
uvicorn
python-multipart
python-dotenv

# Utilities
numpy<2
scipy
tqdm
prettytable
tabulate

# Custom / Others
config     # if it's a local module, should be installed another way
dataset    # same as above, check if it's pip-installable or internal
